User interaction while render process possible

+++ renderrob.py
@@ -33,9 +33,6 @@ class MainWindow(QWidget):
     super().__init__()
     self.window = None
     self.table = None
-    self.number_active_jobs = 0
-    self.job_row_index = 0
-    self.current_job = 0
     self.cache = cache_pb2.RenderRobCache()  # pylint: disable=no-member
     self.state_saver = state_saver.StateSaver()
 
@@ -43,6 +40,11 @@ class MainWindow(QWidget):
     self.process = None
     self.is_saved = True
     self.recent_states = []
+
+    self.active_render_job = None
+    self.green_jobs = []
+    self.yellow_jobs = []
+    self.red_jobs = []
 
   def setup(self) -> None:
     """Provide main function."""
@@ -289,15 +291,26 @@ class MainWindow(QWidget):
           line = line.replace(warning, '')
           color_format.setBackground(QColor(table_utils.COLORS["yellow"]))
           color_format.setForeground(QColor(Qt.black))
-          table_utils.color_row_background(self.table,
-                                           self.job_row_index - 1,
-                                           QColor(table_utils.COLORS["yellow"]))
-          table_utils.set_background_colors(self.table, 987, self.job_row_index)
 
+          self.state_saver.table_to_state(self.table)
+          row_number = state_saver.find_job(
+              self.state_saver.state.render_jobs, self.active_render_job)
+
+          table_utils.color_row_background(self.table,
+                                           row_number,
+                                           QColor(table_utils.COLORS["yellow"]))
         if line.startswith(error) or "blender.crash.txt" in line:
           line = line.replace(error, '')
           color_format.setBackground(QColor(table_utils.COLORS["red"]))
           color_format.setForeground(QColor(table_utils.COLORS["grey_light"]))
+
+          self.state_saver.table_to_state(self.table)
+          row_number = state_saver.find_job(
+              self.state_saver.state.render_jobs, self.active_render_job)
+
+          table_utils.color_row_background(self.table,
+                                           row_number,
+                                           QColor(table_utils.COLORS["red"]))
 
           # Only scroll down if user is at bottom.
         if self.window.textBrowser.verticalScrollBar().value() > (
@@ -320,13 +333,9 @@ class MainWindow(QWidget):
 
     # Find a more elegant way to do this.
     if "Blender quit" in output:
-      self._refresh_progress_bar()
-      self.window.textBrowser.insertPlainText("\n")
       self.window.textBrowser.moveCursor(QTextCursor.End)
 
     if "blender.crash.txt" in output:
-      self._refresh_progress_bar()
-      self.window.textBrowser.insertPlainText("\n")
       self.window.textBrowser.moveCursor(QTextCursor.End)
 
     self.table.blockSignals(False)
@@ -361,10 +370,8 @@ class MainWindow(QWidget):
     if item and isinstance(item, QTableWidgetItem):
       if item.column() == 1:
         table_utils.fix_active_row_path(item, self.state_saver.state.settings.blender_files_path)
-        if not os.path.exists(item.text()) and not os.path.exists(
-                os.path.join(self.state_saver.state.settings.blender_files_path, item.text())):
-          QMessageBox.warning(self, "Warning", "The .blend file does not exist.", QMessageBox.Ok)
-    self.check_table_for_errors()
+    self.table.blockSignals(True)
+    self.set_table_colors()
     self.table.blockSignals(False)
 
   def before_and_after_table_change(self, item: Optional[QTableWidgetItem] = None) -> None:
@@ -381,29 +388,30 @@ class MainWindow(QWidget):
 
   def start_render(self) -> None:
     """Render operator called by the Render button."""
-    self.table.blockSignals(True)
+    self.green_jobs = []
+    self.yellow_jobs = []
+    self.red_jobs = []
+    self.active_render_job = None
     self.window.progressBar.setValue(0)
     self.window.render_button.setEnabled(False)
     self.window.stop_button.setEnabled(True)
-    table_utils.make_read_only_selectable(self.table)
     self.state_saver.table_to_state(self.table)
-    self.job_row_index = 0
-    self.current_job = 0
-    table_utils.reset_all_backgruond_colors(self.table)
-    self.number_active_jobs = self._get_active_jobs_number()
     self._continue_render(0)
     self.table.blockSignals(False)
 
   def stop_render(self) -> None:
     """Interrupt the render operator."""
-    self.process.kill()
-    del self.state_saver.state.render_jobs[:]
+    if self.process:
+      self.process.kill()
+    self.window.progressBar.setValue(0)
+    print_utils.print_info("Render stopped.")
+    self.active_render_job = None
     self.window.stop_button.setEnabled(False)
     self.window.render_button.setEnabled(True)
-    self.window.progressBar.setValue(0)
-    table_utils.make_editable(self.table)
-    print_utils.print_info("Render stopped.")
-    self.window.render_button.setEnabled(True)
+    self.window.textBrowser.moveCursor(QTextCursor.End)
+
+    self.set_table_colors()
+    self.table.blockSignals(False)
 
   def play_job(self) -> int:
     """Open a job in image viewer or Blender Player."""
@@ -447,6 +455,8 @@ class MainWindow(QWidget):
     """Open the output folder of the currently selected job."""
     self.state_saver.table_to_state(self.table)
     current_row = self.table.currentRow()
+    if current_row == -1:
+      return
     snb = shot_name_builder.ShotNameBuilder(
         self.state_saver.state.render_jobs[current_row],
         self.state_saver.state.settings.output_path,
@@ -524,32 +534,88 @@ class MainWindow(QWidget):
     self.table.blockSignals(False)
 
   ######### MAIN WINDOW UTILS ###########
-
-  def _refresh_progress_bar(self) -> None:
-    progress_value = int(100 / self.number_active_jobs) * self.current_job
-    self.window.progressBar.setValue(progress_value)
-
   def _continue_render(self, exit_code: int) -> None:
     self.table.blockSignals(True)
-    # 62097 is the exit code for an interrupted process -> cancelled render.
-    table_utils.set_background_colors(self.table, exit_code, self.job_row_index)
     print_utils.print_info("Continuing render.")
-    if self.state_saver.state.render_jobs:
-      job = self.state_saver.state.render_jobs.pop(0)
-      if not job.active:
-        self.job_row_index += 1
-        self._continue_render(664)
+    # Stop the process if the stop button was pressed.
+    if not self.window.stop_button.isEnabled():
+      return
+
+    # Handle the previous render job and store it in the correct list.
+    if self.active_render_job:
+      if exit_code == 0:
+        self.green_jobs.append(self.active_render_job)
+      elif exit_code == 987:
+        self.yellow_jobs.append(self.active_render_job)
+      elif exit_code in (62097, 11):
+        self.red_jobs.append(self.active_render_job)
       else:
-        print_utils.print_info(f"Starting render of {job.file}")
-        self.job_row_index += 1
-        self.current_job += 1
-        self.render_job(job)
-    else:
+        raise ValueError(f"Exit code {exit_code} not recognized.")
+
+    self.active_render_job = None
+
+    self.state_saver.table_to_state(self.table)
+    self.set_table_colors()
+
+    # Get the next render job.
+    for i, job in enumerate(self.state_saver.state.render_jobs):
+      if job in self.green_jobs or job in self.yellow_jobs or job in self.red_jobs:
+        continue
+      if not job.active:
+        continue
+      self.active_render_job = job
+      table_utils.color_row_background(
+          self.table, i, QColor(table_utils.COLORS["blue_grey_lighter"]))
+      break
+
+    if not self.active_render_job:
       print_utils.print_info("No more render jobs left.")
       self.window.progressBar.setValue(100)
-      table_utils.make_editable(self.table)
       self.window.render_button.setEnabled(True)
+      self.window.stop_button.setEnabled(False)
+    else:
+      all_jobs_count = len([x for x in self.state_saver.state.render_jobs if x.active])
+      done_jobs_count = len(self.green_jobs) + len(self.yellow_jobs) + len(self.red_jobs)
+      self.window.progressBar.setValue(100 * done_jobs_count / all_jobs_count)
+      if self.active_render_job.active:
+        self.render_job(self.active_render_job)
+    self.window.textBrowser.moveCursor(QTextCursor.End)
     self.blockSignals(False)
+
+  def set_table_colors(self):
+    """Set the colors of the table."""
+    self.table.blockSignals(True)
+    temp_active_job = None
+    for i, job in enumerate(self.state_saver.state.render_jobs):
+      if job in self.green_jobs:
+        table_utils.color_row_background(self.table, i, QColor(table_utils.COLORS["green"]))
+      elif job in self.yellow_jobs:
+        table_utils.color_row_background(self.table, i, QColor(table_utils.COLORS["yellow"]))
+      elif job in self.red_jobs:
+        table_utils.color_row_background(self.table, i, QColor(table_utils.COLORS["red"]))
+      elif not job.active:
+        table_utils.color_row_background(self.table, i, QColor(table_utils.COLORS["grey_inactive"]))
+      # Color the active job if a render process is active.
+      elif temp_active_job is None and self.window.stop_button.isEnabled():
+        temp_active_job = job
+        table_utils.color_row_background(
+            self.table, i, QColor(table_utils.COLORS["blue_grey_lighter"]))
+      else:
+        table_utils.color_row_background(self.table, i, QColor(table_utils.COLORS["grey_light"]))
+
+    # Check for duplicates.
+    for row_index in range(self.table.rowCount()):
+      if list(self.state_saver.state.render_jobs).count(
+              self.state_saver.state.render_jobs[row_index]) > 1:
+        table_utils.color_row_background(
+            self.table, row_index, QColor(table_utils.COLORS["yellow"]))
+
+      # Set the background color of the blend path.
+      blend_path_item = self.table.item(row_index, 1)
+      blend_path = blend_path_item.text()
+      if not os.path.exists(blend_path) and not os.path.exists(
+              os.path.join(self.state_saver.state.settings.blender_files_path, blend_path)):
+        blend_path_item.setBackground(QColor(table_utils.COLORS["red"]))
 
   ########## TABLE OPS ############
   def copy_from_cell(self) -> None:
@@ -611,38 +677,6 @@ class MainWindow(QWidget):
     self.process.setArguments(args)
     self.process.readyReadStandardOutput.connect(self._handle_output)
     self.process.start()
-
-  ########### TABLE UTILS #############
-  def check_table_for_errors(self) -> bool:
-    """Check the table for errors."""
-    # Unclean structure, but some other errors are handled in the color_row_background function.
-    # Double occurrences of jobs
-    self.table.blockSignals(True)
-    for row_index in range(self.table.rowCount()):
-      if list(self.state_saver.state.render_jobs).count(
-              self.state_saver.state.render_jobs[row_index]) > 1:
-        table_utils.color_row_background(
-            self.table, row_index, QColor(table_utils.COLORS["yellow"]))
-      else:
-        table_utils.color_row_background(
-            self.table, row_index, QColor(table_utils.COLORS["grey_light"]))
-
-      # Set the background color of the blend path.
-      blend_path_item = self.table.item(row_index, 1)
-      blend_path = blend_path_item.text()
-      if not os.path.exists(blend_path) and not os.path.exists(
-              os.path.join(self.state_saver.state.settings.blender_files_path, blend_path)):
-        blend_path_item.setBackground(QColor(table_utils.COLORS["red"]))
-    self.table.blockSignals(False)
-    # Ignoring animation denoising for now, since # it's deprecated in Blender.
-
-  def _get_active_jobs_number(self) -> int:
-    """Get the number of active jobs."""
-    counter = 0
-    for job in self.state_saver.state.render_jobs:
-      if job.active:
-        counter += 1
-    return counter
 
 
 if __name__ == "__main__":
